Make debugging easier
Get rid of chained method calls (a method returns self so it can be
followed with another method call in the same expression) as this can hide
an exception's origin in the stack trace. This is true for interpreter
raised exceptions like "TypeError: foo() takes 1 positional argument but
2 were given".

Give decorator wrapper functions in astkit.py unique names to help with
exception tracebacks: `def <decorator>_wrapper`.